feat: remove adserver root folder
