+++ AdServerPlugin/Entities/RewardItem.cs
