+++ AdServerPlugin/Enums/AdOrientation.cs
