+++ AdServerPlugin/Exceptions/AdRequestException.cs
