+++ AdServerPlugin/Exceptions/SDKInitException.cs
