+++ AdServerPlugin/Internal/Factories/AdServerFactory.cs
