+++ AdServerPlugin/Internal/Factories/AdViewFactory.cs
