+++ AdServerPlugin/Internal/Factories/DisplayMetricsFactory.cs
