+++ AdServerPlugin/Internal/Factories/InterstitialAdFactory.cs
