+++ AdServerPlugin/Internal/Factories/RewardedAdFactory.cs
