+++ AdServerPlugin/Internal/Helpers/Android/ActivityHelper.cs
