+++ AdServerPlugin/Internal/Helpers/Android/DisplayMetrics.cs
