+++ AdServerPlugin/Internal/Helpers/Android/LayoutHelper.cs
