+++ AdServerPlugin/Internal/Implementations/Android/AdServer.cs
