+++ AdServerPlugin/Internal/Implementations/Android/AdView.cs
