+++ AdServerPlugin/Internal/Implementations/Android/InterstitialAd.cs
