+++ AdServerPlugin/Internal/Implementations/Android/RewardedAd.cs
