+++ AdServerPlugin/Internal/Listeners/AndroidAdServerListener.cs
