+++ AdServerPlugin/Internal/Listeners/AndroidBaseListener.cs
