+++ AdServerPlugin/Internal/interfaces/IAdServer.cs
