+++ AdServerPlugin/Internal/interfaces/IAdView.cs
