+++ AdServerPlugin/Internal/interfaces/ICommonAd.cs
