+++ AdServerPlugin/Internal/interfaces/IInterstitialAd.cs
