+++ AdServerPlugin/Internal/interfaces/IRewardedAd.cs
